refactor(mga): split lineup_repo.py into lineup/ per-pane subpackage
Same shim pattern as PR #758 (classifier_service split). lineup_repo.py
shrinks from 898 → 74 LOC as a thin re-export; the actual implementations
move into a new lineup/ subpackage carved along the 4-pane storyboard:

  lineup/
    filters.py        — LineupFilters + _apply_filters (43 LOC)
    lifecycle.py      — create/get/list/update/hide/accept + classifier
                        writeback (262 LOC)
    throw_pane.py     — set_clip_url + _original + _trim + the throw +
                        widen-source backfill lists (208 LOC)
    landing_pane.py   — set_landing_clip_url + _original + _trim + the
                        landing backfill list (155 LOC)
    micro_panes.py    — set_stand/aim_clip_url + set_stand/aim_screenshot_url
                        + the micro backfill list (185 LOC)
    technique.py      — set_technique + the technique backfill list (76 LOC)
    density.py        — zone_density (58 LOC)

All 25 caller sites use `from app.repositories.game.lineup_repo import X`
and are preserved byte-identically by the shim — no caller churn.

Resolves the High-severity tech-debt item; pairs with PR #758 to close
the file-size discipline initiative. 676/676 backend tests green.
